genai_animation_incubator/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
pydantic>=2.0
pillow>=10.0
